moved structure equations into a single module to avoid 'import *'
